Add Timelens support, to improve video navigation capabilities
This adds support for displaying visual timelines and popup thumbnails
throughout voctoweb's interface, aiming to make video navigation more
efficient, precise, and active. See timelens.io for an explanation of
what this is all about.

Two fields, "timeline_filename" and "thumbnails_filename" are added to
the Event model. When present, the Timelens plugin for MediaElement.js
is activated. Also, Timelens interfaces are added to video listings.