Support slide subdirectories
Some contributors use subdirectories named by year for their uploads; with those slides weren't correctly detected as the directory name didn't start with `slides` but rather with e.g. `2018/slides`. This patch changes the behavior so that the string `slides` can be anywhere in the path and doesn't need to be at the beginning.