Public API: Fix pagination
By default, the api-pagination gem makes it possible to specify the
amount of items per page using the `per_page` query parameter. However,
when a fixed `per_page` value is set during the `paginate`-call, the
parameter is not recognized.

This commit fixes that by specifying a global, default per_page value to
Kaminari and removing local definitions. Possible implication might be a
change of the amount of displayed items in the frontend.

In addition it implements pagination on Public::ConferencesController#index
As this would be a breaking API change, the pagination is only turned on
if either the `page` or `per_page` query parameter is set.