player: fix timelens VTT parsing
The current timelens version generates VTT files with timestamps that
have an hour position instead of using minute values > 60. The version
of timelens.js used in voctoweb does not handle this, causing the
thumbnails of the first minutes to stretch over the entire video.

Fix this by importing the patch proposed in an upstream PR[1].

[1] https://github.com/timelens/timelens.js/pull/3